refactor+feat: create PrecomputedSkeletonMetadata and pkg for sharded
