+++ cloudvolume/datasource/precomputed/skeleton/sharded.py
@@ -1,0 +1,4 @@
+
+
+# class ShardedPrecomputedSkeletonSource(object):
+#   def get(self, segid):
