+++ cloudvolume/datasource/precomputed/skeleton/unsharded.py
@@ -13,14 +13,9 @@ from cloudvolume.exceptions import (
 from cloudvolume.lib import red, Bbox
 from cloudvolume.storage import Storage, SimpleStorage
 
-from .common import cdn_cache_control
+from ..common import cdn_cache_control
 
-from ...skeleton import Skeleton
-
-
-# class ShardedPrecomputedSkeletonSource(object):
-
-#   def get(self, segid):
+from ....skeleton import Skeleton
 
 class UnshardedPrecomputedSkeletonSource(object):
   def __init__(self, meta, cache, config):
@@ -102,86 +97,3 @@ class UnshardedPrecomputedSkeletonSource(object):
       compress='gzip', 
       cache_control=cdn_cache_control(self.config.cdn_cache)
     )
-
-class PrecomputedSkeletonMetadata(object):
-  def __init__(self, meta, cache=None, info=None):
-    self.meta = meta
-    self.cache = cache
-
-    if info:
-      self.info = info
-    else:
-      self.info = self.fetch_info()
-
-  @property
-  def path(self):
-    if 'skeletons' in self.meta.info:
-      return self.meta.info['skeletons']
-    return 'skeletons'
-
-  @property
-  def full_path(self):
-    return self.meta.join(self.meta.cloudpath, self.path)
-
-  def fetch_info(self):
-    info = self.cache.download_json(self.meta.join(self.path, 'info'))
-
-    if not info:
-      return self.default_info()
-    return info
-
-  def refresh_info(self):
-    self.info = self.fetch_info()
-    return self.info
-
-  def commit_info(self):
-    if self.info:
-      self.cache.upload_single(
-        self.meta.join(self.path, 'info'),
-        json.dumps(self.info), 
-        content_type='application/json',
-        compress=False,
-        cache_control='no-cache',
-      )
-
-  def default_info(self):
-    return {
-      '@type': 'neuroglancer_skeletons',
-      'transform': [  
-        1, 0, 0, 0, # identity
-        0, 1, 0, 0,
-        0, 0, 1, 0
-      ],
-      'vertex_attributes': [
-        {
-          "id": "radius",
-          "data_type": "float32",
-          "num_components": 1,
-        }, 
-        {
-          "id": "swc_type",
-          "data_type": "uint8",
-          "num_components": 1,
-        }
-      ],
-      'sharding': None,
-    }
-
-  def is_sharded(self):
-    if 'sharding' not in self.info:
-      return False
-    elif self.info['sharding'] is None:
-      return False
-    else:
-      return True
-
-class PrecomputedSkeletonSource(object):
-  def __new__(cls, meta, cache, config):
-    skel_meta = PrecomputedSkeletonMetadata(meta, cache)
-
-    if skel_meta.is_sharded():
-      raise NotImplementedError()
-
-    return UnshardedPrecomputedSkeletonSource(skel_meta, cache, config)
-
-
